+++ kit/ag-grid/index.js
@@ -17,8 +17,8 @@ import {checkVersion} from '@xh/hoist/utils/js/VersionUtils';
 export let AgGridReact = null;
 export let agGridVersion = null;
 
-const MIN_VERSION = '24.1.0';
-const MAX_VERSION = '24.*.*';
+const MIN_VERSION = '25.0.1';
+const MAX_VERSION = '25.*.*';
 
 /**
  * Expose application versions of ag-Grid to Hoist.