+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 
 ## v38.0.0-SNAPSHOT - unreleased
 
+
+### ⚙️ Technical
+* We have removed the experimental flags `useTransactions`, and `deltaSort` from `GridModel`.
+The former has been the default behavior for Hoist for several releases, and the latter is
+obsolete.
+
+### 💥 Breaking Changes
+* Hoist grids now require ag-Grid v25.0.1 or higher - if your app uses ag-Grid, update your
+ag-Grid dependency in your app's `package.json` file.
+
+
+
 ### 🎁 New Features
 
 [Commit Log](https://github.com/xh/hoist-react/compare/v37.2.0...develop)